VirtualInertia

Documentation for VirtualInertia.

NetworkDynamics.ODEEdge

MetaGraphs.get_prop

MetaGraphs.set_prop!

VirtualInertia.DroopControl

VirtualInertia.PT1Source

VirtualInertia.PerfectSource

VirtualInertia._group_states

VirtualInertia.bustype

VirtualInertia.describe_edges

VirtualInertia.describe_nodes

VirtualInertia.import_system

VirtualInertia.import_system

VirtualInertia.subscript

NetworkDynamics.ODEEdge — Type

Conventions: (Anti-)symmetric lines:

inputs: u_src_r, u_src_i, u_dst_r, u_dst_i

outputs: i_r, i_i

current direction is defined from src to dst

dst node will receive -i_r, -i_i

Asymmetric lines:

inputs: u_src_r, u_src_i, u_dst_r, u_dst_i

outputs: i_src_r, i_src_i, i_dst_r, i_dst_i

not yet implemented. might by tricky with fidutial...

MetaGraphs.get_prop — Method

get_prop(g, keys::Iterable, prop::Symbol)

Get same property prop with different values vals for differet keys.

MetaGraphs.set_prop! — Method

set_prop!(g, keys::Iterable, prop::Symbol, vals::Iterable)

Set same property prop with different values vals for differet identifiers keys.

VirtualInertia.DroopControl — Method

DroopControl(; params...)

Return block for droop control outer ocntrol.

VirtualInertia.PT1Source — Method

PT1Source(; params...)

Create Schiffer Voltage source which follows angle directly but amplitude with lag.

VirtualInertia.PerfectSource — Method

PerfectSource()

Perfect Voltage source which follows the reference directly.

VirtualInertia._group_states — Method

Find complex states with _r and _i suffix

VirtualInertia.bustype — Method

bustype(i::Int)
bustype(s::String)
bustype(s::Symbol)

Helper to convert the different representations of bus types into eachother. Base representation is Symbol:

:load   == bustype(1) == bustype("PQ") == bustype(:PQ)
:gen    == bustype(2) == bustype("PV") == bustype(:PV)
:syncon == bustype(3) == bustype("Ref") == bustype(:Ref)

VirtualInertia.describe_edges — Method

describe_nodes(g::MetaGraph; firstcols=Vector{String}())

Returns DataFrame with all the edge meta data.

VirtualInertia.describe_nodes — Method

describe_nodes(g::MetaGraph; firstcols=Vector{String}())

Returns DataFrame with all the node meta data.

VirtualInertia.import_system — Method

import_system(sym::Symbol; kwargs...)::MetaGraph

Main entry point to load the systems. New systems should overload this function. Known implementations

import_system(:rtsgmlc): loads the GMLC update for the rts96

Those functions return a MetaGraph which has properties attached to the Nodes/Edges.

Graph properties:

Pbase : Base power for PU

optional:

NodeProps : tuple of node property names which should appear first in describe functions

EdgeProps : tuple of edge property names which should appear first in describe functions

Node properties:

P : active power in PU

Q : reactive power in PU

Vbase : base voltage in kV

Vm : voltage magnitude in PU

optional:

inertia : inertia in MJ/MW (see H in Wikipedia page)

damping : damping factor γ for swing equations

timeconstant : time cosntant τ for dynamic loads

x, y : position of Bus for plotting purposes

Va : voltage angle in rad

Edge properties:

rating : short term emergency rating in PU

R : resistance in PU

X : reactance in PU

VirtualInertia.import_system — Method

import_system(:rtsgmlc)

Import the RTS-GMLC system as a MetaGraph.

VirtualInertia.subscript — Method

subscript(s, i)

Append symbol or string s with a integer subscript.